refactor: duplicate hook
