+++ packages/plugins/RedPacket/src/SiteAdaptor/hooks/useRedPacketCover.ts
